% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4EstimateWithCA.R
\name{C4EstimateWithCA}
\alias{C4EstimateWithCA}
\title{Estimate photosynthesis parameters for C4 species using Sharkey's fitting procedure}
\usage{
C4EstimateWithCA(
  ACi,
  alpha1 = 0.15,
  x = 0.4,
  CaBreakL = 10,
  CaBreakH = 45,
  startp = c(45, 300, 3, 20, 120, 100)
)
}
\arguments{
\item{ACi}{Gas exchange measurement from LI-6800 equipment. It is a
dataframe iput. Ci with the unit of ppm.}

\item{alpha1}{The fraction of O2 evolution occurring in the bundle sheath.
Unless you have enough information, input it as the 0.15.}

\item{x}{the fraction of total electron transport that are confined to be
used for the PEP regeneration out of J, which is the total electron transport.}

\item{CaBreakL}{Higher bound of Ci below which A is thought to be controled by
Rubisco Carboxylation (Ac). Start with 10.}

\item{CaBreakH}{Lower bound of Ci above which A is thought to be controled by
RuBP regeneration (Aj). Start with 50. If the estimation results showed
"inadmissible fits", change the CaBreakL and CaBreakH until "inadmissible fits"
disappear.}

\item{startp}{A vector that gives the start points for the estimation
(c(Vcmax,J,Rd,gm,Vpmax,KCA))}
}
\value{
This package will return a dataframe that contains the following values
(c(Vcmax,J,Rd,gm,Vpmax and KCA)). You can try with c(30, 150, 3, 10, 50, 80).

Parameter at leaf temperature:      A vector (c(Vcmax,J,Rd,gm,Vpmax and
KCA)) returns the estimation parameters at leaf temperature.

Parameter at 25°C:                  A vector (c(Vcmax,J,Rd,gm,Vpmax and
KCA)) returns the estimation parameters at leaf temperature.

Objective:                          The final objective value based on
the estimation results.

Convergence:                        An integer code. 0 indicates
successful convergence.

Message:	                          A character string giving any
additional information returned by the optimizer, or NULL. For details, see PORT
documentation.

Iterations:	                        Number of iterations performed.

Evaluations:	                      Number of objective function and
gradient function evaluations.
}
\description{
Using the gas exchange measurement (A_Ci curve), C4 photosynthesis model with
carbonic anhydrase reaction and Sharkey et al. (2007) fitting processure to do
nonlinear curve fitting (using nlminb package) for estimating photosynthesis
parameters (Vcmax,J,Rd,gm and Vpmax) for C4 species. Make sure to load the "stats"
package before intstalling and using the "C4Estimation" package.
}
