% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntensiveACiFit-package.R
\docType{package}
\name{IntensiveACiFit-package}
\alias{IntensiveACiFit}
\alias{IntensiveACiFit-package}
\title{IntensiveACiFit: C4 ACi curve fit tools for LI-6800}
\description{
This is a revised code from Zhou (2019) et al specially used for LI-6800.
}
\author{
\strong{Maintainer}: Jiedong Zhu \email{zhujiedong@yeah.net} (\href{https://orcid.org/0000-0001-7651-2684}{ORCID})

}
\keyword{internal}
